% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefko3-package.R
\docType{package}
\name{lefko3-package}
\alias{lefko3}
\alias{lefko3-package}
\title{Historical and Ahistorical Population Projection Matrix Analysis}
\description{
This package creates population matrix projection models (MPMs)
for use in population ecological analyses. It presents a complete working
environment for the construction and analysis of ALL kinds of MPMs and IPMs,
including age, stage, and age-by-stage versions. Its specialty is the
estimation of historical MPMs, which are 2d matrices comprising 3 monitoring
occasions (2 time steps or periods) of demographic information. The package
constructs both function-based and raw MPMs for both standard ahistorical
(i.e. 2 occasions, 1 time step) and historical analyses, has functions for
complex density-dependent and independent, and stochastic and cyclical,
projections, and also includes the automatic calculation of quality control
metrics throughout every step of analysis. It also includes powerful
functions to standardize demographic datasets.
}
\details{
The lefko3 package provides seven categories of functions:

1. Data transformation and handling functions

2. Functions determining population characteristics from vertical data

3. Model building and selection

4. Matrix / integral projection model creation functions

5. Population dynamics analysis and projection functions

6. Functions describing, summarizing, or visualizing MPMs and derived
structures

7. Extra functions used to illustrate core theory and ideas.

lefko3 also includes example datasets complete with sample code.
}
\references{
Shefferson, R.P., J. Ehrlen, and S. Kurokawa. 2021. 
\emph{lefko3}: analyzing individual history through size-classified matrix 
population models. \emph{Methods in Ecology and Evolution} 12(2): 378-382.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://revolutionarydemography.com/projects.html}
}

}
\author{
\strong{Maintainer}: Richard P. Shefferson \email{cdorm@g.ecc.u-tokyo.ac.jp} (\href{https://orcid.org/0000-0002-5234-3131}{ORCID})

Authors:
\itemize{
  \item Johan Ehrlen (\href{https://orcid.org/0000-0001-8539-8967}{ORCID})
}


Richard P. Shefferson <cdorm@g.ecc.u-tokyo.ac.jp>

Johan Ehrlén
}
