% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefko3-package.R
\docType{package}
\name{lefko3}
\alias{lefko3}
\title{Historical and Ahistorical Population Projection Matrix Analysis}
\description{
This package creates population matrix projection models (MPMs)
for use in population ecological analyses. Its specialty is the estimation
of historical MPMs, which are 2-dimensional matrices comprising 3 monitoring
occasions (2 time steps or periods) of demographic information. The package
constructs both function-based and raw MPMs for both standard ahistorical
(i.e. 2 occasions, 1 period) and historical analyses, and can also produce
age-by-stage MPMs and IPMs. It also includes powerful functions to
standardize demographic datasets.
}
\details{
The lefko3 package provides seven categories of functions:

1. Data transformation and handling functions

2. Functions determining population characteristics from vertical data

3. Model building and selection

4. Matrix / integral projection model creation functions

5. Population dynamics analysis functions

6. Functions describing, summarizing, or visualizing MPMs and derived
structures

7. Extra functions used to illustrate core theory and ideas.

lefko3 also includes example datasets complete with sample code.
}
\references{
Shefferson, R.P., J. Ehrlen, and S. Kurokawa. 2021. 
\emph{lefko3}: analyzing individual history through size-classified matrix 
population models. \emph{Methods in Ecology and Evolution} 12(2): 378-382.
}
\author{
Richard P. Shefferson <cdorm@g.ecc.u-tokyo.ac.jp>

Johan Ehrlén
}
