% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sf_skeleton}
\alias{sf_skeleton}
\title{Create Skeleton Stageframe}
\usage{
sf_skeleton(stages, standard = TRUE)
}
\arguments{
\item{stages}{The number of stages, as an integer.}

\item{standard}{A logical value indicating whether to create a standard
\code{stageframe} object (\code{TRUE}, the default), or a reassessed
\code{stageframe} object as created by function \code{mpm_create()}
(\code{FALSE}).}
}
\value{
A data frame of class \code{stageframe}.
}
\description{
Function \code{sf_skeleton()} creates a skeleton \code{stageframe} object.
}
