% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdyn.R
\name{summary.lefkoLTRE}
\alias{summary.lefkoLTRE}
\title{Summarize lefkoLTRE Objects}
\usage{
\method{summary}{lefkoLTRE}(object, ...)
}
\arguments{
\item{object}{A \code{lefkoLTRE} object.}

\item{...}{Other parameters currently not utilized.}
}
\value{
A list of data frames. In all cases, the first data frame is one
showing the positive, negative, and total contributions of elements in
each LTRE contribution matrix. If not a SNA-LTRE, then there are an
additional two (if deterministic) or four (if stochastic) data frames. If
deterministic, then \code{hist_det} is a data frame showing the summed LTRE
contributions for all 16 kinds of historical transition per matrix, with each
column corresponding to each A matrix in order, followed by all summed
positive and all summed negative contributions. Object \code{ahist_det} is a
data frame showing the summed LTRE contributions for all four kinds of
ahistorical transition per matrix, with order as before, followed by summed
positive and summed negative contributions. If stochastic, then
\code{hist_mean} and \code{hist_sd} are the summed LTRE contributions for the
mean vital rates and variability in vital rates, respectively, according to
all 16 historical transition types, followed by summed positive and negative
contributions, and \code{ahist_mean} and \code{ahist_sd} are the equivalent
ahistorical versions. The output for the SNA-LTRE also includes the
logs of the deterministic lambda estimated through function \code{ltre3()}.
}
\description{
Function \code{summary.lefkoLTRE()} summarizes \code{lefkoLTRE} objects.
Particularly, it breaks down LTRE contributions by the kind of ahistorical
and, if applicable, historical transition.
}
\examples{
data(lathyrus)

sizevector <- c(0, 100, 13, 127, 3730, 3800, 0)
stagevector <- c("Sd", "Sdl", "VSm", "Sm", "VLa", "Flo", "Dorm")
repvector <- c(0, 0, 0, 0, 0, 1, 0)
obsvector <- c(0, 1, 1, 1, 1, 1, 0)
matvector <- c(0, 0, 1, 1, 1, 1, 1)
immvector <- c(1, 1, 0, 0, 0, 0, 0)
propvector <- c(1, 0, 0, 0, 0, 0, 0)
indataset <- c(0, 1, 1, 1, 1, 1, 1)
binvec <- c(0, 100, 11, 103, 3500, 3800, 0.5)

lathframe <- sf_create(sizes = sizevector, stagenames = stagevector,
  repstatus = repvector, obsstatus = obsvector, matstatus = matvector,
  immstatus = immvector, indataset = indataset, binhalfwidth = binvec,
  propstatus = propvector)

lathvert <- verticalize3(lathyrus, noyears = 4, firstyear = 1988,
  patchidcol = "SUBPLOT", individcol = "GENET", blocksize = 9,
  juvcol = "Seedling1988", sizeacol = "Volume88", repstracol = "FCODE88",
  fecacol = "Intactseed88", deadacol = "Dead1988",
  nonobsacol = "Dormant1988", stageassign = lathframe, stagesize = "sizea",
  censorcol = "Missing1988", censorkeep = NA, censor = TRUE)

lathsupp3 <- supplemental(stage3 = c("Sd", "Sd", "Sdl", "Sdl", "Sd", "Sdl", "mat"),
  stage2 = c("Sd", "Sd", "Sd", "Sd", "rep", "rep", "Sdl"),
  stage1 = c("Sd", "rep", "Sd", "rep", "npr", "npr", "Sd"),
  eststage3 = c(NA, NA, NA, NA, NA, NA, "mat"),
  eststage2 = c(NA, NA, NA, NA, NA, NA, "Sdl"),
  eststage1 = c(NA, NA, NA, NA, NA, NA, "NotAlive"),
  givenrate = c(0.345, 0.345, 0.054, 0.054, NA, NA, NA),
  multiplier = c(NA, NA, NA, NA, 0.345, 0.054, NA),
  type = c(1, 1, 1, 1, 3, 3, 1), type_t12 = c(1, 2, 1, 2, 1, 1, 1),
  stageframe = lathframe, historical = TRUE)

lathsupp2 <- supplemental(stage3 = c("Sd", "Sdl", "Sd", "Sdl"), 
  stage2 = c("Sd", "Sd", "rep", "rep"),
  givenrate = c(0.345, 0.054, NA, NA),
  multiplier = c(NA, NA, 0.345, 0.054),
  type = c(1, 1, 3, 3), stageframe = lathframe, historical = FALSE)
  
ehrlen3 <- rlefko3(data = lathvert, stageframe = lathframe, year = "all", 
  stages = c("stage3", "stage2", "stage1"), supplement = lathsupp3,
  yearcol = "year2", indivcol = "individ")

ehrlen2 <- rlefko2(data = lathvert, stageframe = lathframe, year = "all",
  stages = c("stage3", "stage2"), supplement = lathsupp2,
  yearcol = "year2", indivcol = "individ")

ehrlen3ltre <- ltre3(ehrlen3)
summary(ehrlen3ltre)

}
