% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sup_skeleton}
\alias{sup_skeleton}
\title{Create A Supplement Skeleton Data Frame}
\usage{
sup_skeleton(rows = NULL)
}
\arguments{
\item{rows}{An integer giving the number of rows to include.}
}
\value{
A data frame with the format of a supplement, of class
\code{lefkoSD}.
}
\description{
Create A Supplement Skeleton Data Frame
}
\examples{
new_supp <- sup_skeleton(3)
new_supp

}
