% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelselection.R
\name{create_pm}
\alias{create_pm}
\title{Creates a Skeleton Paramnames Object for Use in Function-based Modeling}
\usage{
create_pm()
}
\value{
A three column data frame, of which the first describes the
parameters in reasonably plain English, the second gives the name of the
parameter within the MPM generating functions, and the third is to be
edited with the names of the variables as they appear in the models.
}
\description{
Creates a simple skeleton paramnames object that can be entered as input in
functions \code{\link{flefko2}()}, \code{link{flefko3}()}, and
\code{\link{aflefko2}()}.
}
\examples{
our_pm <- create_pm()
our_pm

}
